Seperate View Filtering Example and Store Filtering Example
